Specimen Page Written
- Created fundamental structure of the specimen page, containing the
following information about the specimen in order: the name, latin
name, a photograph taken from museum, introductory description, 3D
(STL) View, 3D surface markings of interest (written as a paragraph
beneath the viewer), Cross-sectional (DICOM) View, Cross-sectional
points of interest (written as a paragraph beneath the viewer). [Note
all of the information in this case has been directly applied to the
.php file and will be taken from an updated .xml file in the next
commit]
- Redesigned navbar for the specimen page. Navbar’s are often intrusive
when trying to learn and focus on the page content. In this case the
navbar has been faded/blurred out and comes back into focus on hover. A
delay has also been introduced so it does not return to its resting
blurred state immediately (Note as the blur effect is a new in css,
only certain browsers such as safari support it).
- The new specimen page can be accessed from the “collections”
(landing) page. By clicking on any of the explore specimen buttons.
Eventually each button will lead to a page specific to the selected
specimen.
- Unity Player is no longer on a separate page and works inline in the
page. The branding and fullscreen buttons have been commented out. The
fullscreen button did not seem to function correctly and the logo was
getting in the way of the description below. Everything else about the
Unity Player is functioning correctly and it works perfectly.
- Dicom viewer not working, and I have created an image substitute for
the meantime. Something to work on integrating into the page next.
- Small visual changes and tweaks.