Better navigation, xml & php support
- Pages renamed. The landing page has been renamed to collections, and
the starting page renamed to welcome.
- Unecessary files deleted (certain html files were not being used.
- The specimen page now works fully with the xml document. All the
names, descriptions etc are pulled from the xml file. One bug is that
the stll viewer is still fixed on the hedgehog model and does not
change with the information.
- The specimen page now has a javascript button at the end taking the
user back to the collection he/she was looking at previously.
- The new collections page (or old landing page) is in the process of a
revamp (some bugs present and photos need updating). The user can
choose between the five offical categories and an other section.
Clicking on one of the categories opens an in-window popup, allowing
the user to select what he or she wants to learn more about. The
advantage of this 'popup' is that the information is all loaded from
the beginning shortening waiting times.
- Other small tweaks and additions include the use of font-awesome
buttons, a folder has been created under api.